+++ tests/io_performance_test.py
@@ -67,7 +67,7 @@ def get_mean_cycle_count(io_dtype, capfd):
                           (torch.float32, torch.float16)])
 @helpers.printCapfdOnExit
 @unittest.mock.patch.dict("os.environ", helpers.disableAllModels())
-@helpers.overridePoptorchLogLevel("INFO")
+@helpers.overridePoptorchLogLevel("DEBUG")
 def test_compare_io_performance(capfd, io_dtype1, io_dtype2):
     cycle_count_1 = get_mean_cycle_count(io_dtype1, capfd)
     cycle_count_2 = get_mean_cycle_count(io_dtype2, capfd)